Add a calls-to-sym query term, disabled context-menu support
This adds "calls-to-sym" as a counterpart to "calls-to" which takes an
identifier.  I added a commented out line in context-menu.js that can
be switched to as desired for some quick A/B debugging.  For the case
I was interested in, this was not helpful though, as the first problem
is that the callers that exist as exposed by a use's contextsym are
locals which end up without a kind.  Symbol visibility would be next.