Handle multiple symbols per pretty for overloads/etc.
Without these fixes, we would only register the first symbol with a
given pretty which falls down in the face of overloaded methods or
variations due to variations between platforms, also possibly for
template functions (when supported).

It would be possible to detect that we are dealing with multiple
symbols and try and introduce a suffix to the pretty, but I think
it's better to defer that fix to having the crossref deal with that
and introduce more human-understandable and potentially consistent
suffixes that can live on the symbol.  Specifically, doing this
correctly requires distinguishing between overloads which have
distinct definition points versus platform/template variations which
should all share the same (spelling) definition point/loc.