Display the symbol for overloads.
When we added the symbol information we didn't expose it to the UI, so
this exposes the (raw) symbol.  Of course, showing the raw symbol is
not great UX, so it would be nice to do a pass to fill in the jumprefs
for symbols which we don't already have the crossref info for.  (There
are cases like the node limit where we won't have done the crossref
lookup, and arguably it seems preferable to avoid making it the
problem of code which might only have the symbol id to need to do the
lookup for the pretty identifier.)