Bug 1417586 - Handle preproced IPDL, determine C++ locations from objdir-files lookup.
This updates us to use https://github.com/mozsearch/ipdl_parser/pull/12
so we can parse PContent, and with the remainder of the changes here,
correctly do the C++ semantic linkage as well.

Previously we used a predicted path, but this doesn't work for preprocessed
files which can't be unified by the platform merge step for trees like
mozilla-central.