Improve resiliency for bad SourceRange values.
From the m-c try run I believe we were emitting empty ranges which the
parser was assuming would not be empty.  So this changes us to not emit
empty ranges but also to accept them if we encounter them, as there's
definitely no need to crash in these cases and it's a likely failure
type that could recur.