Bug 1829273 - Fix crash when processing NSS
It appears that sometimes lines has a length of zero, and so the
lines.remove(0) crashes. This updates the guard to protect against
that case too. I have no idea what this code is doing though so
maybe there's a better fix, but this should hopefully at least
unblock the indexer.