rust: Prefer scip index if available.
With the latest mozilla-config and scip changes, this gives me a working
mozilla-central index.

There are some regressions that I noticed that might be worth addressing
first:

 * No stdlib symbols: If you have `use std::io::Write;`, clicking
   `Write` won't find all references to it. Not super-surprising given
   we have a bunch of extra code to deal with the rust stdlib in save
   analysis, but maybe we can do something about it.

 * No module definitions: If you have `mod foo;`, clicking `foo` won't
   show the "Go to definition" button to get you to the file.

 * #[no_mangle] functions keep the rust-module-namespaced symbol names.
   I'm not sure how easy to address is this without upstream fixes,
   we'll see.

Other than that, it seems to be working, and maybe we can even enable it
by default if those regressions get worked on or if we consider they're
not such a huge deal.