Improved query arg parsing for graph-format for debugging.
The root problem turned out to be that graphviz was crashing and leaving
a core file, but in the process of being able to explicitly output the
raw graphviz input for graph-format, I improved the argument handling
logic so that different terms can have different priorities for their
impact on merging arguments.  This was necessary because the graph-render
step always gets applied after the user terms.