Add "debug" query param option to emit tracing logs from query.
The axum handler method quite properly indicated that it was not safe
for us to try and hold an EnteredSpan across our async calls, so we
move the LoggedSpan to just be a span which is not entered, and then we
have to make sure to use `instrument` so that inside our async waits we
enter the span.

This gets us the core of running the graph for the query, but when I
attempted to enter the span for the sync portions of building the graph
I ran into a problem being able to fully drop the span.  And there are
lexical issues for defining a more bounded scope, so I think for right
now we're just missing potentially useful data.  test_checka_insta is
also therefore perhaps partially regressed, and this may have broken
the crossref use of the logged span too?  Going to dig into that at
least briefly.