Hier diagram improvements by rank control for clusters.
We place all the inferred "input" nodes in a cluster at the top of the
cluster so that nodes with edges to these nodes have a rank that places
them above the cluster rather than potentially having their ranks
overlap and end up being placed next to the cluster.

This also includes a minor fix to the prefix/suffix logic for avoiding
weird edges for constructors to include a delimiter so that we don't
kick edges between methods "ExecWorker" and "ExecWorkerOnMainThread" up
to the cluster's parent but instead leave them in the cluster.