Bug 1842632 - Don't allow continuing on error during setup
For m-c based repos, setup failure can result in the git repo
being left on the master branch rather than the target branch (e.g.
mozilla-beta). So continuing with the indexer in such a case
can result in the "mozilla-beta" tree on searchfox actually
showing the contents from mozilla-central, which is a pretty
surprising/undesirable fallback behaviour. It's better to just
let the indexer abort completely.