Add "inheritance-diagram" for overrides only (not classes), layout control
A quick addition of an "inheritance-diagram" query option which will
transitively traverse "overrides" relationships given any method.  This
obviously would want to work on classes as well, but that probably wants
cmd_traverse to finally be overhauled to operate like
`cmd_crossref_expand` which makes it much cleaner dealing with
differing sets of edge traversals, especially as it relates to
borrowing the underlying symbols.

"graph-layout" is also added to support "neato" and "fdp".  A minor
hack is in place to change the default "overlap" of "true" for "neato"
to "prism" because neato otherwise makes everything overlap.  These
modes definitely want to alter the hierarchical layout behavior since
we lose our clusters.  Visually, something like using a metaball
background coloration likely works.  We can use SVG filters to perform
a gaussian blur and then fix that up to accomplish that.  This would
require we create a solidly colored alternate version of the SVG to
blur that has the same underlying layout.

The most "in-domain" and flexible means of accomplishing the alternate
SVG rep would likely be to use gvpr.  So the pipeline might look like:
- Generate the normal dot layout.  Add a synthetic attribute which can
  have any name that's not a real attribute, as it should be passed
  through by the layout engine, it's just that the output enginer
  currently won't propagate values.  (Although maybe a reasonable
  enhancement might be to teach the SVG output how to do that?)
  - In the graphviz forums the HTML style "data-FOO" has been proposed
    although there's been some indication some naive graphviz grammars
    can't handle the dashes, but that may not matter for us.
  - "data-blobcolor" would be an example of us propagating the explicit
    color we want (which should be a legal dot color).
- Pass the layout through gvpr, having it:
  - Rewrite nodes so that they are filled with the "data-blobcolor" and
    do not have black edges or text or anything like that.
  - Nuke the edges, as they are not helpful.
- Render the original pre-rendered dot to SVG.  (We can do this earlier
  too.)  I think we use `neato -n2` for this.
- Render the transformed metaball dot to SVG.  This also gets to be
  `neato -n2` I believe.
- Munge the metaball SVG into the normal layout SVG, adding the SVG
  effect transforms so that our stack looks right.

One thing I'm not proposing above but which we could try is to also add
text labels for the hierarchy clusters.  This seems like it would
complicate layout a bit and might be more pragmatic to put a legend on
the page and/or hit test the background metaball layer in order to map
the hit color back to the underlying class and then do some emphasis in
the legend and/or add some kind of hover information near the cursor.
The hover information could use the metaball layer to also try and find
a spot where there is no metaball.  (One could even have a non-quantized
version of the metaball layer that could leverage gradient descent to
more quickly find the edges, but it's probably more sane to do something
like stepping in the 8 cardinal directions in a breadth-first loop to
find the closest location there's a big enough gap.  That said, it could
still make sense to have an alternate rendering of the SVG with the blur
with a much more aggressive blur and quantization so that the stepping
doesn't need to deal with false positives.  Visually we want much more
gooey metaballs than we would want for label positioning.)