Symdex seems to work correctly.
Note that the tree-sitter-mozcpp unfortunately doesn't know about the XPIDL
NS_NO_VTABLE macro used by the checked-in XPIDL generated code, so we end up
with some false claims of nsISupports definitions that seem to be due to
tree-sitter's error handling.  Since the generated files aren't checked in
to mozilla-central this probably isn't a huge problem, but we can certainly
also add that as a known macro to the grammar.  Not dealing with that now
because that tends to be a slippery slope and I expect there will be other
fixes that may be necessary.