Address graphviz-rust escaping issue, add test.
graphviz-rust has an `esc` macro that's a bit misleading as it does not
actually do any escaping of the value, it just marks that the value
should be wrapped in double-quotes, which is an interesting choice but
probably one that won't be changed anytime soon.

I've manually inspected and updated the various flows to graphviz-rust
reps, but it's complicated by the 2-stage rendering approach of compile
(which populates the HierarchicalLayoutActions which can contain ids)
followed by render.  Also that we need to sanitize the port identifiers
to a stronger standard than node identifiers because they end up as
part of the node identifiers.

Longer term it would probably be preferable to add some newtype
wrappers so that as soon as we move into graphviz space we have
explicit types that explicitly state whether they've undergone escaping
and potentially have helper methods that produce the appropriate
graphviz-rust types if it reads acceptably.