Address an identifier bounds-range panic for needles longer than ids.
This was revealed from fresh testing of `calls-from` where we ended up
with a wacky mangled template identifier while trying to derive
hierachy.  Examples on the `id` we were subscripting we crashed on:
- `mozilla::UniquePtr<type-parameter-0-0[],`
- `js::WrappedPtrOperations<JS::Value,`

These needles were long and surfaced a logic error in our attempt to
derive a slice for the suffix.  we now guard the suffix logic on there
actually being a suffix, as we probably should have done before.  This
was likely due to a naive translation of the python code for this where
I did not compensate for the fact that Python normalizes slices but
rust does not.

The mangled identifier seems to be due to:
- We 100% are putting pretty identifiers with spaces in the
  `identifiers` due to identifiers like `Foo<Arg0, Arg1>::method`.
- The lookup method does a naive `split(' ')`, and so if we were
  looking up that full identifier, the id would 100% end up shorter
  than the needle.  We should be using `rsplit_once` here because
  that's our format and avoids us needing to deal with the whitespace
  in the pretty at all.

I'll fix this in the next patch, but note that the hierarchy logic does
extremely naive splitting on "::" regardless of the presence of
template arguments (which can include "::" themselves), so that will
still be badly broken without a fix.  And that fix will need to deal
with annoying template arguments like `1 << sizeof(blah) * 8` which
means the parsing needs to be more clever than following `<` pairs.