Bug 1490144 - Build fake symbols in every cases, not only for locals
Commit 7fb450aa introduced support for locals in JS that don't have a
proper symbol definition in the SCIP index.

This also happens in Java and Kotlin for at least:
 * packages, which aren't really defined in any particular document
 * imported classes, interfaces etc, which are defined in dependencies

Therefore this extends the solution from 7fb450aa of creating fake
symbols in every case if no actual SCIP symbol was found.

This is done by extracting the analysis of symbol descriptors into a
separate function and re-using it when an unknown symbol is first met.