Bug 1490144 - Don't let scip-java emit inverse relationshups
By default scip-java emits some inverse relationships and doesn't make
the distinctinction between the upward and downward links.

This has the effect that when an interface A extends an interface B,
Searchfox didn't actually know whether A extended B or if B extended A.

By chance it actually got it right most of the time, because of another
issue where relationships with symbols declared later are ignored.
See github.com/mozsearch/mozsearch/pull/667#discussion_r1394283303

The test case I add here verifies that the list of supers for A is
empty. When the aforementioned bug is fixed, we could replace
crossref_lookup with crossref_expand to verify that B appears in A's
subclasses.