clang-plugin: add binding_slots and slot_owner from C++ annotations
When analysing classes, functions and data members, this looks for
annotations of the form:
```cpp
__attribute__((annotate("[binding_to|bound_as]", "[cpp|jvm]", "[class|method|getter|setter|const]", "<symbol>")))
```
and adds the relevant entries to the output structured analysis'
slot_owner and binding_slots.

This can be set manually in the source code by the author, but is also
meant to be automatically added by previous passes which would
recognize patterns corresponding to automatic binding generators.