Extend binding system for non-IDL-based bindings
The existing binding slot system assumes that bindings are IDL-based,
with an interface defined in an IDL files and implementations of that
interface in various languages.
This mainly shows in the search page and context menu:
 * When searching for a symbol which has a slot_owner the search page
   displays the slot_owner under the IDL heading.
 * When searching for a symbol which has binding_slots, the linked
   symbols are displayed under the Definitions heading.

This adds an owner_lang field to BindingSlotProps. This field can be
set to BindingOwnerLang::IDL in which case we retain the existing
behavior. Or it can be set to another implementation language, in which
case:
 * When searching for a symbol which has a slot_owner the search page
   displays the slot_owner under the Definition heading, as this is the
   definition of the function in the original language, which hold the
   interesting code.
 * When searching for a symbol which has binding_slots, the linked
   symbols are displayed under the new Bindings heading. Searching for
   this symbol will reveal users in the foreign language.