clang-plugin: add @WrapForJNI-based binding analysis
This searches the source code for patterns which match what the
@WrapForJNI Java annotation and code generator generates, and adds the
relevant annotations to the bindings/bound items to be picked up by the
code added in the parent commit.

The patterns we search for are roughly documented in the comments above
the 4 new toplevel functions in BindingOperations.cpp.

Any deviation from the expected patterns result in an early return and
no annotation is added.