Ensure all slot_owners have a matching binding slot
We already ensured that each binding slot in binding_slots had a
matching slot_owner, with the same properties (kind, lang) but in the
opposite direction.

This makes sure the opposite is also true and that every set slot_owner
has a matching binding slot with the same properties.

I extracted the binding properties from StructuredBindingSlotInfo into
a separate struct to make it more obvious that those should match both
ways. This extra struct is marked with #[serde(flatten)] so serialized
json should not change.

In other words we ensure that:
```rust
for (symbol, structured) in &meta_table {
  for binding_slot in structured.binding_slot {
    let slotted = meta_table.get(binding_slot.sym);
    let owner_slot = slotted.owner_slot.unwrap();
    assert_eq!(owner_slot.sym, symbol);
    assert_eq!(owner_slot.props, binding_slot.props);
  }
  if let Some(owner_slot) = structured.owner_slot {
    let owner = meta_table.get(owner_slot.sym);
    let binding_slot = owner.binding_slots.iter().find(|&slot| slot.sym == symbol).unwrap();
    assert_eq!(binding_slot.sym, symbol);
    assert_eq!(binding_slot.props, owner_slot.props);
  }
}
```