clang-plugin: add JNICALL-based binding analysis
This looks for functions whose name match the pattern that the Java
runtime searches by default* when a method is marked “native”, and adds
the relevant annotation to be picked up by the grandparent commit.

Any deviation from the expected pattern results in an early return and
no annotation is added.

*This default behavior is often overridden by calling
JNIEnv::RegisterNatives.