Downgrade all ontology warnings to info.
We don't need emails about violations of this logic.  As noted
elsewhere, our parsing of pretty identifiers to understand types for
pointers is known to be a bad idea, so the medium term plan is to have
clang emit structured metadata for us that avoids us needing to parse
things.  (And/or we can parse the mangled idnentifiers.)