Fix sticky highlight.
The changes made for diagram hovering regressed the inline JS using the
`javascript:` URL scheme.  This patch changes us to no longer doing
that, instead adding an explicit event listener with a lambda.  This
change should make it possible for us to add some CSP.

The one awkwardness is that for simplicity I left the DOM the same, so
we have an `a` tag with an href of `#` and we preventDefault() it and
also stopPropagation() it because we explicitly have the sticky
highlight hide the context menu (as it was already doing).  This could
likely be cleaned up.