Change automated email subjects to not misleadingly thread.
gmail ignores text in square braces at the start of subjects for its
thread inference algorithm and only provides the subject of the first
message in a thread by default, at least when we use square braces.
This makes it harder to triage automated warning and failure emails
because it requires additional clicks in the gmail UI to figure out
if the release1 jobs have had problems or not.  (One must click into
the thread, then click on the "show details" down-arrow.)

This change removes the square braces in an attempt to make sure that
config1 / release1 failures get their own threads in the gmail UI.