Bug 1468445 - implement field-layout and related infrastructure
The UI is not amazing on this and the context menu option makes my muscle
memory a bit irritable, so I've introduced a new "semanticInfo" feature
gate.

This introduces a new symbol tree table result type which also introdues
some new very limited markup support.  Our markup isn't horrible but could
be improved.  I think we probably want to put the offset and length in
their own columns like the fancy branch prototype had done (in React)
using a colgroup.  But I think that also likely wants to happen only once
another format-symbols mode is introduced.

Something this absolutely reveals is that the structured "variants"
mechanism where we hash on content is causing duplicates.  Additionally,
this raises the issue that we really need to the root structured rep to
be a composite of all variants, not just one arbitrary one.  Otherwise
the diagram functionality and anything else that is trying to treat the
root structured object as canonical is potentially going to be misleading.