Bug 1779939 - update git-cinnabar to 0.6.3
With the android repo merge into mozilla-central, git fetch on central
was throwing.  This may potentially have been a transient issue with
the server, but since there wasn't a reason to be on an old version of
cinnabar, I updated us to the most recent release before checking if
things were now okay.

I'm leaving us using a specific release for reproducability reasons,
but honestly if I'd thought to check if there was a "release" branch
(there is, it's up-to-date), I probably would have used that.  That
might be the next thing to do if things ever break again.