Use mode=hier for neato per graphviz forum suggestion
In a graphviz forum post it was mentioned that neato has a hier mode
(see https://graphviz.org/docs/attrs/mode/) which attempts a top-down
layout.  A notable piece of feedback for the neato/sfdp layouts versus
the default dot layout was that they were much more stressful to look
at, even if they provided potentially more useful density.  This was
trivial to quickly hack in and A/B test thanks to previous efforts in
this space.

In particular, this mode seems quite nice for larger inheritance
diagrams, but one notable thing was self-edges for XPCOM interfaces
where the pretty identifier for the XPCOM interface are the same for
the C++ class overlapping adjacent nodes.  To this end, I also
increased the "sep" margin (https://graphviz.org/docs/attrs/sep/).