Fix overridenBy diagram arrow direction to be consistent.
The overrides arrow direction was right, but the overriddenBy direction
was just copied and pasted without being reversed.  This meant that while
the following 2 queries should effectively look the same, they did not:
- `inheritance-diagram:'mozilla::dom::cache::ReadStream::Controllable::CloseStream'`
- `inheritance-diagram:'mozilla::dom::cache::ReadStream::Inner::CloseStream'`

A slipstream change is to also make us traverse overriddenBy for the
"callees" diagram.  This makes
`calls-from:'mozilla::dom::cache::StreamControl::CloseAllReadStreams' depth:4`
more useful which is how I noticed the problem when bringing up the
inheritance diagram for `Controllable::CloseStream`.