Generate synthetic symbol names for main functions.
Main functions will have a tendency to collide, so use a location-based
scheme with an "MF_" prefix.  This choice is arbitrary and something
we can change at any time.

This came up in the previous commit where a test case was showing symbol
over a main symbol in terms of its subsystem; the specific test had
previously been picking up the main symbol from a different file, and
it's likely this test case would be unstable going forward without this
change.