Support additional levels of jumpref traversal exposure in context menu
From the DOM meetup week, :peterv pointed out that we don't have an
option to go to the "recv" method when we're looking at a send symbol.
The reason for this was that we only would potentially load jumprefs at
a distance of one from the initial symbol referenced in a source record,
but the pretty symbol identifier was at a distance of 2 because we need
to traverse the owning binding slot to the parent (distance 1), then
traverse down the "recv" binding slot (distance 2).

This patch generalizes the logic in crossref_converter.rs and its use in
format.rs to support the send/recv use-case plus something that had been
bothering me when dealing with XPCOM interfaces where I would be looking
at a call to an XPCOM interface but I really want to go to the singular
override implementation of that interface.  The latter only requires a
distance of 1, of course.

This patch also contains changes to context-menu.js to support
generating the additional context menu items, but only a direct link to
their definition (and not any declaration).

The net change in behavior here is expected to be:
- IPDL
  - When clicking on `actor->SendFoo` we get a new "Go to definition of
    IDL C++ recv Blah::RecvFoo".
  - When clicking on the `async Foo` in the ipdl file, nothing changes.
- XPIDL where there is a single override/impl.
  - When clicking on `nsIPointer->Foo()` we get a new "Go to definition
    of Sole Override ImplClass::Foo" in a cpp file.
  - When clicking on `Foo` in the XPIDL file we get a new "Go to
    definition of Sole Override ImplClass::Foo" as the second entry
    after the existing "Go to definition of IDL C++ method nsIBar::Foo".
- XPIDL when there are two overrides:
  - Same as above but we have "Override" instead of "Sole Override".
- XPIDL when there are three or more overrides:
  - No changes.  Although honestly there is a case where for Service
    Workers we have 4 impls and I'm a bit tempted to bump the number to
    4 because of that, but I expect there are potentially some complex
    cases where that might be too much, so I figure we start with 2.

Note that I did reformat `format.rs` and it induced some spurious
changes but they are trivial enough and the code boring enough (and
potential adoption of hyperblame in the future will moot them enough)
that I'm just leaving that as-is.