Build fix: `cargo update` run to get time 0.3.36 update to fix nightly build failure.
As noted on the thread for the indexer failure, we failed to update because the
new rustc nightly does not time 0.3.34.  So I did a `cargo update` and then did
a fixup for tree-sitter-rust which had performed a divergent upgrade to 0.22.x
which I manually downgraded to 0.20.9 to match everything else.