Bug 1904474 - Explicitly specify docker platform
On Apple Silicon macbooks, it defaults to running the aarch64
ubuntu image, which then fails during docker image build because
the scripts try to install amd64 things (graphviz and much more).
I did explore the option of having the scripts choose binaries
of the correct architecture but it was much more involved and
this seems like a more straightforward solution.