Bug 1904373 - pipeline-server ulimit should be aware of mmap data sizes
Initially we used a hard-coded ulimit of 24G based on observed values
on config1 out of a combination of erring on the side of safety and
because I worried the shell scripting would be a headache and likely to
break.  But it turns out the arithmetic context offered by bash via
`$(( $a + $b ))` is easy to use in a way I honestly did not expect.

This patch translates the original rationale for arriving at 24G to be
done based on invocations of du.  This works as expected on config4
(where I ran an updated version of the script) as well as config1
(where I ran the shell scripts manually).