Bug 1858453 - Index mozsearch and mozsearch-mozilla
The changes here are to add a "build-mozsearch-repo" / "serve-mozsearch-repo"
pair which will use the newly introduced "just-mozsearch.config" config file
from the mozsearch-mozilla tree.

I also update the newrepo docs to cite use of lz4 since we've started doing
that since wubkat.  Note that we have not transitioned existing tarballs,
and that if we clean that up we might also want to take the win from having
our tarballs not actually have a working tree checked out.  (All the
information is in the `.git` dir and we can just check things out after we
untar.)

Note that most changes are in the paired mozsearch-mozilla PR.