Bug 1906580 - Use consistent padding for .content box
Instead of applying wider padding for non-source-listing cases,
add padding/margin for items inside it when necessary,
so that the breadcrumbs is positioned in the same place across pages.

Spacing for the following are kept by adding extra padding/margin:
  * directory listing table
  * field layout table

Spacing for the top page and the settings page are reduced, but
they should already have sufficient padding:

Spacing for other queries (call graph, use graph) are also reduced,
but they have spacing in its own.