Bug 1911131 - Forward all relevant jumpref data to SYM_INFO.
When populating SYM_INFO, only the first symbol of merged symbols was
output to HTML. This led to missing “Go to definition” entries in some
cases involving macros or templates.

For templates, see the new test files templates6.h/templates6.cpp and
the corresponding webtest.

For macros this is especially visible in the atom_list.h use-case,
except in our test repo because atom_list.h doesn't go through the
merge-analyses step.