Bug 1918801 - Switch to tree-sitter-kotlin-ng, upgrading tree-sitter
As a new tree-sitter grammar, kotlin-ng is based off the most recent
tree-sitter, so it's necessary to upgrade.  Most grammars we use are
maintained with the exception of tree-sitter-mozcpp, but we only use
that for the shelved hyperblame effort, so it's not an active concern.

I've made some minor changes to the language match block factory idiom
that net out to just deferring setting the parser and compiling the
queries until after the match block.  I initially tried to just pass a
LanguageFn out, but for the tree-sitter-cpp case we did want access to
the language in the match block to extra some node metadata, so I
updated both uses of the idiom to look like this.

With the change to tree-sitter-kotlin-ng the queries did need a slight
change as described in the bug because there's no "body" field, so we
have to match on the node type instead.  The changes I made to support
this in the previous patch are now used.