Bug 1918797 - emit referenced but not-defined SCIP symbols
For references to JDK/Kotlin/AndroidX classes, without this fix we have
no structured records for them, which makes it harder for the ontology
logic to work cleanly.  This makes us emit all the structured records
we learned about during SCIP indexing but did not actually emit.  We
emit them under a synthetic analysis file named after the SCIP index
file in the root of the analysis directory.  We give them a bogus
Location which should allow merging to work.

We change the crossref process to process structured records before
we check for and read in the source file.  We need the source file to
extract snippets, but obviously that's not necessary or valid to do in
this synthetic case.

This works as expected for superclasses like `S_jvm_androidx/work/Worker#`
where we end up with structured information because we also see ocurrences
in the source.  But this does not currently do anything for overrides like
`S_jvm_androidx/work/Worker#doWork().` where there are no occurrences,
just structured metadata references.  This will be addressed in the next
patch.