Bug 1918797 - emit structured records for all external symbols
For the superclass and overridden method cases where we learn about
symbols that are not otherwise seen as definitions or ocurrences,
generate fake structured records for those as well.  We emit these,
plus the previous fake structured records, as "external" rather than
"impl" for the purposes of clarity.  This shouldn't be user visible
since we should not know how to emit jumps for them, but it is a good
descriptor if it does end up user-exposed.