Bug 1918797 - emit crossref entries for external symbols
The previous change to include the structured external symbols in the
meta map was not sufficient.  We also need to make sure to have mapping
entries for forward and reverse pretty/symbol mappings, as well as
creating a top-level "table" entry so that we will actually emit the
structured record.