Bug 1918797 - Make Function3::invoke runnable work with JVM constructor
From a slack thread, kotlin::Function3::invoke was brought up as
interesting.  This is a case where we want the normal C++ runnable
"constructor" heuristic but for Kotlin the constructor is "<init>", not
the class-name, so we adapt to that.  I use a bit of a hack based on
JVM symbols starting with "S_jvm_" but it would obviously be nice to
just have that info available on the structured record.