Bug 1922888 - Fix IndexConsumer::getContext(SourceLocation) when Ctxt->Decl is null.
Bug 1856762 added support for traversing the implicit declarations
created for lambdas. When traversing a lambda, a new context is pushed
to the AutoSetContext stack to set the VisitImplicit flag. This context
has Decl == nullptr because the lambda is not a NamedDecl.

I modified IndexConsumer::getContext(Decl *D) to support that case, but
forgot to modify getContext(SourceLocation Loc) to support it too. In
the added test case, the non-catch-all captures triggered a crash.