Bug 1922408 - java/kotlin/gradle build updates to compensate for openjdk-21
I think I ended up updating more than was strictly required here.  The
core issue is that the change from v19 to v21 required a change in
`build.gradle.kts` where we had hardcoded v19 as
`languageVersion.set(JavaLanguageVersion.of(19))`.  However, the gradle
error message was very misleading because it seemed like the problem
was that gradle was confused about what was going on in
'/usr/lib/jvm/openjdk-21', but it was just a noisy warning that could
be ignored.  Newer versions of gradle no longer report the situation
like it's a problem which is why I've kept the gradle upgrade here.

I performed the gradle upgrade using
`./gradlew wrapper --gradle-version latest` in tests/files per the
instructions at
https://docs.gradle.org/current/userguide/gradle_wrapper.html#sec:upgrading_wrapper

I also updated the `com.sourcegraph:semanticdb-*` referencing lines to
match what we are using in m-c right now and bumped
"org.jetbrains.kotlin.jvm" as part of my technique of pushing random
buttons until things worked.