Bug 1922408 - make upgraded `cargo insta review` happy
`cargo insta review --workspace-root=/vagrant/tests/` stopped working
because it started doing a `cargo metadata` check which was not happy
about the lack of a Cargo.toml there.  So I changed us to use a more
specific workspace path (tests/tests/checks) and stuck a fake
Cargo.toml in there.

I've also updated the "insta" crate in our tools/Cargo.toml, but note
that this breakage we're addressing is actually because a fresh
provisioning does `cargo install cargo-insta` without a specific
version specified and so any fresh provisioning would have this
problem.