Bug 1924227 - Extract root block device from nvme list rather than hardcoding it
In some cases the EBS root volume may not be the previously hardcoded
`/dev/nvme0n1` so we should extract it via jq.  The idiom we've been
using in cases like this is to run the command that gives us a JSON
blob once into an INFO var and then use separate jq invocations that use
`<<< "$FOO_INFO"` to then pick the relevant data out of there rather
than issue multiple commands which might change dynamically.  I've done
that here.

This successfully ran on both provisioning jobs I just kicked off,
although the hardcoded values would have worked too, but I did consult
the logs and we are using the new logic.

Permuted to deal with github problems.