Bug 1453506 - Forward type-dependent news to template call locations.
This provides a way to make type-dependent statements appear as-if they
were located at the template call site.

During the analysis of the template, this first registers all type-dependent
function call locations in ForwardedTemplateLocations (we don't have more
information at this point).

Then when analyzing the template instantiations, we link the actual
type-resolved statements to the template instantiation.

Then when emitting the analysis data for a template call, we also (recursively)
emit all the analysis data for the forwarded items at the call site.