Bug 1931631 - add crossref check showing erroneous use by class
This use should be attributed to the constructor, not the class; the
next patch in the stack will correct that.

This check could have been a filter-analysis check but the crossref
check provides slightly better context for the change since it shows
a "line" excerpt and better serves the high level goal of properly
attributing all uses of the symbol.  Also, it's nice to have some
incidental additional structured record test coverage for fields.