Bug 1931631 - the constructor should be the context of field initializations
Pre-patch the code explicitly is skipping the constructor decl because
of the documented behavior of `getContext(Decl)`.  It seems likely the
original intent matches what I've changed it to: we explicitly use the
constructor decl as the context without relying on the implicit stack.

The changes to the stackartholder / thing / outercat callees are all
cases where the field uses were being attributed to the class and
therefore shown as callees.  (Which I suppose demonstrates one of
the mixed bag of test checks that produce large amounts of output;
there is a potential on initial checkin to notice and catch bugs if
carefully looking at all the output, but frequently the not directly
relevant data may just be eyeballed.  And then when changes like this
are made the diffs are both informative but potentially noisy.)