Emit confidence information to HTML
This adds a `data-confidences` attribute next to the existing
`data-symbols` attribute. This attribute contains a JSON-encoded array
of stringified ConfidenceLevels.

Contrary to `confidence` field of the AnalysisSource source class, we
always emit confidences for all symbols, even if they are all
`Concrete`. This keeps the code in format.rs simpler.