Bug 1933325 - Run the AnalyzeDependent phase on all template declarations with a definition.
The code previously checked that a class template declaration was
“canonical” to traverse its specialization and instantiations. This
canonical declaration could be a forward declaration. In those cases,
the GatherDependent phase would not record any locations to analyze.
→ The Maybe test in ForwardingTemplates.cpp would fail with the new
forward declaration.

For function templates however, while the canonical declaration may be
in-class and the definition out-of-line, traversing the inline
declaration will correctly traverse the out-of-line definition too.
→ The new emplace_out_of_line test would fail if we didn't traverse the
canonical declaration.

This performs the GatherDependent phase for all templates declarations
that are either canonical or have their own body, then performs the
AnalyzeDependent phase on all related instantiations.