Bug 1933325 - Unify ForwardedTemplateLocations insertions for CallExprs.
Instead of implementing this in the various Visit* functions, we can
centralize most of them in VisitCallExpr. Only the location of the
interesting token in the callee varies.

This also moves the remaining of the code out of the isInterestingLoc
guards, as we also want to track calls made through the STL for
instance.

CXXNewExpr has to stay separate because it is not a CallExpr.